Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

I think you're right, this is already covered by -Wall.

I had noticed that adding this didn't cause any compilation errors, but pushed it up to Gerrit in the hope that it would prevent future issues from cropping up, and that we just happened to be clean now. But as you point out, this is already turned on by -Wall, so this CL doesn't add anything new.

This wouldn't catch the issue in the bug you mentioned since that promotes the enum value to an int when its returned and saved to a local. Then it's okay to compare the int to (almost) any enum value.

Patch-set: 1
Reviewer: Gerrit User 1260470 <1260470@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>