Update patch set 7
Patch Set 7:

Only full committers OR the CL owner if they have tryjob access or the CL is approved are allowed to trigger CQ runs.
CL is missing the approving vote on Verified label.
Committers in this project are members of:
 * https://chrome-infra-auth.appspot.com/auth/groups/project-chromeos-committers
Tryjob access is additionally granted to members of:
 * https://chrome-infra-auth.appspot.com/auth/groups/googlers

Patch-set: 7
Tag: autogenerated:cq:full-run