host_command: Change host command return value to enum ec_status
If the host command handler callback function returns an int, it's easy
to accidentally mix up the enum ec_error_list and enum ec_status types.
The host commands always expect an enum ec_status type, so we change the
return value to be of that explicit type. Compilation will then fail if
you accidentally try to return an enum ec_error_list value.

Note that as written this only covers files needed by nocturne_fp. If
the approach is acceptable, we would need to do a mass change.

BRANCH=nocturne
BUG=chromium:1004831
TEST=make BOARD=nocturne_fp -j

Change-Id: Id93df9387ac53d016a1594dba86c6642babbfd1e
Signed-off-by: Tom Hughes <tomhughes@chromium.org>