EFS: Use major & minor version for rollback protection
If we use the commit count for rollback protection, we can't go back
to the previous version even if an older version does not contain any
vulnerability.

This change makes EFS use the major and minor version instead.
The advantage is that we can control when to move the rollback
protection version forward.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/119329144,chromium:998135
BRANCH=none
TEST=Verify RW version is corrected printed on console.

Change-Id: Ic49803799edcd355f7e1588d6445a538aa429e4f