nvmem: do not advance page index if it points to an empty page
When new objects are added to NVMEM, sometimes it comes to the state
where the object ends at the very top of the last page.
nvmem_get_data() in this case advances the master list_index value,
but does not yet allocate the new page.

This patch detects this condition when start_new_flash_page() is
invoked and uses the page pointed at by the incremented list_index.

BRANCH=cr50, cr50-mp
BUG=b:140139297
TEST=who hatch devices survived 1500 iterations of suspend_stress_test

Change-Id: I1de90a7758e7edebacbb0f894f8fec916e7a0adb
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>