fpsensor: Send positive match salt and validation value on enrollment.
On enrollment success, generate new positive_match_salt, compute
positive_match_secret and validation_value, and send positive_match_salt
as part of encrypted blob and validation_value as a field in the metadata.
The positive_match_salt is used to derive positive_match_secret and
is different from the encryption salt for encrypting the templates.
The positive_match_salt needs to be sent to biod and stored with templates
because it needs to be re-uploaded to FPMCU the next time the user logs in.
The validation value needs to be sent to biod so that biod can validate
the positive_match_secret at the next match.

BRANCH=nocturne
BUG=chromium:927095
TEST=make -j buildall
TEST=tested enrollment, matching and multifinger on DUT nocturne

Change-Id: If4e3d3a7f2cbf777bfbddb947442277bf9969ea4
Signed-off-by: Yicheng Li <yichengli@chromium.org>