util/comm-i2c: Add switch to specify I2C bus
Currently, devices running CrOS EC that aren't actually the embedded
controller (such as touchpad or fingerprint MCUs) can only be contacted
over I2C if they are in the device tree. To avoid having to recompile
the Kernel, Coreboot, or `ectool`, the `--i2c_bus` switch allows
`ectool` to be used for testing as-is.

BRANCH=none
BUG=none
TEST=Checked various commands (hello, version, inventory...) with
     `--i2c_bus=7`, connecting to a MAX32660. Verified that a
     contradictory `--interface` switch is rejected. Checked that
     invalid bus numbers (≥32) are rejected.

Change-Id: I92f3307bbbdf88978b9f8271610a3ae222279767
Signed-off-by: Harry Cutts <hcutts@chromium.org>