Makefile: Ignore -Waddress-of-packed-member for GCC
GCC 9 enabled the -Waddress-of-packed-member warning by default, which
causes the build to fail. This is already ignored for Clang, so do the
same for GCC. Tidy up the organization of the flags while we're at it.

Conflicts:
	Makefile.toolchain: UBSAN not defined.

BUG=none
BRANCH=none
TEST=make buildall -j

Change-Id: I8796b268503c9fef2be9336ee01beaf5309024f6
Signed-off-by: Jacob Garber <jgarber1@ualberta.ca>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1826289
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
(cherry picked from commit 344f19b1c150d0a3fc5f88b45735ab5910c16783)