ec_commands: Obsolete v0 and declare v1 of host command EC_CMD_POWER_INFO
v0 of this command is no longer used (no handler is defined), so it is
removed.  The new host command version gives the AP the ability to
dynamically change its power settings based on the battery status,
A/C, USB-PD, and the static configuration of the system and its
battery.

Also modified ectool to support v1 of this command and deprecate
support for v0.

BUG=b:139840435
BRANCH=master
TEST=buildall

Change-Id: I516d838581400314e626fd679ca1c5afd5e34ff6
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>