Update patch set 1
Patch Set 1:

i don't have a kodama board here to verify that we actually end
up with sane code with this patch.  could someone working on
kodama do a quick check?

test:
patch in
https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1670653
at which point, your build should fail.
then patch in this one and see if you still have a functional board.
the affected area is the USB-PD state machine, so if PD charging 
still works, we're probably OK.

Patch-set: 1