Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Can you please incorporate my comments on https://crrev.com/c/1761736 ?
> > 
> > Also, it would be nice if this were split into ISH5.4 chip variant, and tglrvp_ish board in two separate CLs.
> 
> I will handle your comments on https://crrev.com/c/1761736.
> 
> I prefer to not split the patch since it will be much clear for an initial enabling patch, people who reference this patch will complete know how ish 5.4 for tgl rvp is enabled in one stop. For following feature enhancements or bug fix, surely will split accordingly.

I agree with Jack on this. We want to have 2 different CLs here, one for the chip code and one for the board. They can be in a CL stack so it is clear that they go together, but they shouldn't be submitted as one CL

Patch-set: 2
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>