cr50: nvmem: fix list_index after erased header at page end
Before this CL, if the last 4 bytes of the last page were not written,
after reading those bytes to check if they contain a header, the mt was
rewinded to the end of the previous container, but list_index stayed
pointing to the next page. Which could lead to skipping pages when
creating the next container. This CL fixes it.

BRANCH=none
BUG=b:140139297
TEST=running test_nvmem with CL:1828184 applied doesn't lead to error 14

Change-Id: Icc8d1b8eb5cfc0f0230706ed386c2ef9b4936cae
Signed-off-by: Andrey Pronin <apronin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1830372
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
(cherry picked from commit 56f611ffdc7a52f192f898cb15e1eb7e9e940ba2)