Update patch set 2
Patch Set 2:

(4 comments)

The board.h change needs to be split out into a separate CL that would be Helios specific. The remaining 3 files are related to a change to the driver. However, I have a comment about the need to guard this feature with a config option.

Patch-set: 2
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>