common: Add random number generator
This change adds a random number generator and two
distributions: uniform and normal. This is needed
to generate random noise in sensor bias to prevent
fingerprinting the device once bias is exposed to
the application layer.

This change includes the addition of %f format
specifier for ccprintf. Note that because not all
chips support double precision, and GCC promotes
single precision floats to doubles by default when
using varargs, a pointer to the float must be used.

BUG=chromium:966253
BRANCH=None
TEST=buildall with new unit tests
Change-Id: I919ae7533fd42b0394de66aa0585e58343a662cc
Signed-off-by: Yuval Peress <peress@chromium.org>