common: Add sensor stillness detector
This change adds a stillness detector for 3d sensors.
This will be needed to filter sensor readings when
calibrating later.

This change includes the addition of %f format
specifier for ccprintf. Note that because not all
chips support double precision, and GCC promotes
single precision floats to doubles by default when
using varargs, a pointer to the float must be used.

BUG=b:138303429
BRANCH=None
TEST=buildall with new unit tests
Change-Id: I919ae7533fd42b0394de66aa0585e58343a662cc
Signed-off-by: Yuval Peress <peress@chromium.org>