trembyle: turn on FRS
As the PD specification indicates, we must ignore the FRS
signal until
  1) we are sink
  2) we have an explicit contract
  3) we have received a Sink Capabilities from the source
  4) the sink capabilities must indicate FRS enabled
  5) and we must be able to handle the required current
Once these requirements are met, we enable FRS listening.

Trembyle will use TCPC_REG_POWER_CTRL_FRS_ENABLE to control
FRS listening by the TCPC and FASTSW_CTL GPIO to allow fast
switching.

When the TCPC detects the FRS Signal, if it is listening for
it, will cause the PPC to switch from sink to source and
sends an alert to the EC.

BUG=b:1385999955
BRANCH=none
TEST=make buildall -j

Change-Id: I987446967ecc000f34f159d4373a62d6a7c492d1
Signed-off-by: Denis Brockus <dbrockus@chromium.org>