Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I think we should get rid of -1 from tablet_mode.
> > 
> > I assume -1 is needed to pass the check in tablet_set_mode so that we can reach hook_notify(HOOK_TABLET_MODE_CHANGE) and host_set_single_event(EC_HOST_EVENT_MODE_CHANGE) on start-up.
> > 
> > I think we should make the host and the hook handlers think the default state is clamshell (tablet_mode=0) so that they are synced on reset.
> 
> this cause failure here:
> in motion_lid.c, current_mode is set to tablet_get_mode()
> motion_lid_set_tablet_mode() {
> ......
> /* Only change tablet mode if we're sure. */
> 		if (current_mode != new_mode) {
> 
> ......}

If motion_lid thinks it's tablet (according to the current lid angle), it'll switch to tablet mode if tablet_get_mode returns 0. If the lid angle is in clamshell zone, it'll be skipped (because tablet_get_mode would return 0).

I don't think we need to fix this part of the code. What we have to fix is the code where some initialization is done based on what tablet_get_mode returns.

Patch-set: 1