WIP-retimer: Port changes under CONFIG_BB_RETIMER to TCPMv2
To configure the burnside bridge retimer, current DP pin mode,
cc state and the type of the cable is required.

For the same, implemented a board level function that returns the
current DP pin mode and added a function that returns the type of
cable inaccordance to the cable vdo response.

Also added a new version to USB_PD_CONTROL host command, to return
the DP mode, cc_state and the cable type

BUG=b:141971044
BRANCH=None
TEST=To get correct CC state, pin mode and cable type
Pin mode: USB:0x0 (No DP)
          DP cable:0x4 (Mode:C)
          USBC dock:0x8 (Mode:D)

Change-Id: If87ae6b77e5fa2ceaa22319dfa2d2c802460edfa
Signed-off-by: Ayushee <ayushee.shah@intel.com>