sysjump: handle panic_data changes between RO and RW
we need to handle the case where the size of the panic_data struct
changes between the RO and RW images.  when panic_data is present, the
jump_data is placed just below panic_data in memory.  the RO image
needs to compute the address of the jump_data using the same math used
by the RO image.

BRANCH=none
BUG=b:142031466
TEST=sysjump no longer powers down the AP

Change-Id: I06ffc71c45e11d3695272d0a9ba8323e34cf1096
Signed-off-by: Caveh Jalali <caveh@chromium.org>