Update patch set 2
Patch Set 2: Verified+1

'verified' bit got wiped out by a rebase.

Does this still need anything?

IMO it is fine to keep this corner case unaddressed in the test, as it is unlikely that we'd modify the size of the var space such that it would result in a whole number of test var sizes fitting. But even if this happens, there would be no problem with the actual code, just the test would fail. And that would be a good time to address that corner case.

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1