Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> I believe Vijay provided a possible fix for the error you were seeing. Are you still observing the compile time errors?
> 
> Kindly add aaboagye@ once this CL is ready for review.

Hey Prashant,
There is a continuous port reset (Logs in b:141971044) when a type-C to DP connector is attached. Looks like the soft reset is because the cable is non emark hence, causing a  transmission failure. But, according to USB PD 3.0 Sec 6.4.4.3.1 : In this case a Discover Identity Command request sent to SOP' Shall Not cause a Soft Reset if a GoodCRC Message response is not returned since this can indicate a non-PD Capable cable.

Patch-set: 2