it83xx/spi: add spi slave function
Add the spi slave function which is required to
communicate with the EC when the CPU is the ARM
processor.

This spi slave function is used to receive data
by FIFO mode. When the data transfer is completed,
the interrupt signal will generate, and the data
will pass to common layer to handle request.

BUG=none
BRANCH=none
TEST=none

Change-Id: I65dc42033dc8cc4ac86be966d55b47d2f892f659
Signed-off-by: tim <tim2.lin@ite.corp-partner.google.com>