tablet_mode: expose uninitialized value for table_mode
The lid angle module needs to know that the tablet mode is uninitialized
so it can properly set it to either tablet or notebook mode when the EC
is initializing. Without returning the uninitialized value of -1, one of
the modes will not be initialized correctly.

All other uses of get_tablet_mode treat this as a boolean, and
implicitly convert -1 to 1 in all other cases.

BRANCH=none
BUG=b:141494453
TEST=put arcada at 200 degrees and see that tablet mode it initialized
correctly.

Change-Id: I4e23d3ba149f9add84f9667a5af676803cf50da5
Signed-off-by: Jett Rink <jettrink@chromium.org>