battery: Move usb_get_battery_soc in USB to get_battery_soc in battery
For other functionality unrelated to USB, it seemed unnecessary to have to
bring in USB-related functionality just to get battery SoC.  So this was
moved to common/battery.s instead.  Also needed to fix up a few boards
which use Type-C but have no battery (samus_pd and fizz) and one test as
well.

BUG=none
BRANCH=none
TEST=buildall

Change-Id: I13d9887884cf828d75e3d18854476811cc82138c
Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>