g: update eraseflashinfo to match new policies
With moving to the new Cr50 RO only images which have a certain bit in
their header set and are signed by the prod key will be allowed to
erase the INFO1 space.

This patch modifies the eraseflashinfo console command to report
failure in case the bit in the header is not present.

To maintain backwards compatibility with RO 0.0.10 while it is
available. this check is not compiled in into images with CR50_DEV=1.

This will allow dev images built from ToT to erase INFO1 space until
transition to the new RO is completed.

BRANCH=cr50-mp
BUG=b:74100307,b:187438971
TEST=make BOARD=cr50

Change-Id: I44d0e78c7257be153f7994ef20fc03893e8b8109
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1856819
Reviewed-by: Andrey Pronin <apronin@chromium.org>
(cherry picked from commit 10346db1a08568f1e4bbfde67d05d014bbf42110)
Signed-off-by: Mary Ruthven <mruthven@chromium.org>