g: update eraseflashinfo to match new policies
With moving to the new Cr50 RO only images which have a certain bit in
their header set and are signed by the prod key will be allowed to
erase the INFO1 space.

This patch modifies the eraseflashinfo console command to report
failure in case the bit in the header is not present, and includes the
command only if CR50_DEV is not defined at the compilation time.

BRANCH=cr50-mp
BUG=b:74100307
TEST=verified that attempts to erase fail with RO 0.0.10.

Change-Id: Id8abea7f35676f55fbd06f2a48a83cb271e210f6
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>