system.c: reset statics for testing
this adds a function to reset the state information generated by
system_common_pre_init() for testing scenarios that need to call it
multiple times.  on the EC, main memory (.data + .bss) is
reinitialized across sysjumps, so this happens automatically, but we
can't really do that from unit tests.  so, add a function to reset the
relevant static variables to emulate main memory getting
reinitialized.

BRANCH=none
BUG=b:142031466
TEST=make buildall passes

Change-Id: I1f65902c21ab6fc17c32388795cfef19c84d8cc8
Signed-off-by: Caveh Jalali <caveh@chromium.org>