Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Does it make sense to redefine all  or at least crypto-test extension commands then as vendor commands? There are some dependencies for commands like EXTENSION_FW_UPGRADE (gsctool, trunks, something else?) and EXTENSION_POST_RESET  (gsctool), but I don't see same for crypto-test extensions - where is code which use it?

The only drawback for this would be that Python based tests would stop working on older images, but maybe this is not a big deal as this is purely internal, so this should be fine.

Another issue is gsctool using EXTENSION_FW_UPGRADE and EXTENSION_POST_RESET - gsctool would have to be modified to allow communicating with both versions before these are dropped.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>