Update patch set 1
Patch Set 1: Code-Review+2

The code is better than it was with Jett's change but I did find areas that still fail. I was working on it until Jett came back to see if I could solve it but disabling this is a great idea until it is fixed.

Patch-set: 1
Reviewer: Gerrit User 1346479 <1346479@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2