Update patch set 3
Patch Set 3:

yeah, this flake has been killing a lot of builds for other people's CLs and has been making the CQ painful.  getting a fix like this is fine when there isn't an obvious answer, or when the team wants to debate a different direction.  feel free to follow up with other changes now that chromeos-ec isn't blowing up unrelated builds.

Patch-set: 3