fpsensor: Copy BEP and libfp source files from private repo
These source files have been approved to be released publicly (see
http://b/137848573#comment20). There were no source changes in the move,
except for small changes to the Makefiles to adjust for the move.

For future "git blame"s refer to the history of the files in the private
repo if you need to access the full history.

BRANCH=none
BUG=b:137848573
TEST=make buildall -j

Cq-Depend: chrome-internal:2005127
Change-Id: I1a809ac5a4d95637840ee47712e9c80a40a445c5
Signed-off-by: Tom Hughes <tomhughes@chromium.org>