fingerprint: Restrict md and rw console commands
This CL ultimately restricts access to the md and rw console
commands when the system is locked, when system_is_locked().
It does this by enabling and configuring the restricted console
command feature for nocturne_fp and hatch_fp.

This sets up the framework for more console commands to use the
restricted flag for fingerprint, but right now, only the md and rw
command use the restricted console command flag.

The md and rw commands enabled the restricted flag with balhh.

BRANCH=nocturne
BRANCH=nami
BRANCH=hatch
BUG=b:142559996, b:142505927
TEST=none yet

Change-Id: I1202a118b836ff2a2acc986f6f3b0f6e07b36894
Signed-off-by: Craig Hesling <hesling@chromium.org>