fingerprint: Restrict md and rw console commands
This CL ultimately restricts access to the md and rw console
commands when the system is locked, when system_is_locked().
It does this by enabling and configuring the restricted console
command feature for nocturne_fp and hatch_fp.

This sets up the framework for more console commands to use the
restricted flag for fingerprint, but right now, only the md and rw
commands use the restricted console command flag.

The restricted flag for md and rw was introduced in crrev.com/c/1868430.

BRANCH=nocturne
BRANCH=nami
BRANCH=hatch
BUG=b:142559996, b:142505927
TEST=# Test on nocturne_fp
     make BOARD=nocturne_fp
     scp build/nocturne_fp/ec.bin dut1:~/
     ssh dut1 flash_fp_mcu ./ec.bin
     # From FPMCU UART console
     version
     # Ensure that version is custom based on workstation
     help
     # Ensure that md and rw are not prefixed with "-"
     help list
     # Ensure that a "Flags" column exists and that
     # md and rw both have a 1 in the column.
     md 0x0
     # Ensure that this successfully reads the word
     rw 0x24000000
     # Ensure that this successfully reads the word
     syslock
     md 0x0
     # Ensure that this fails with "Access Denied"
     rw 0x24000000
     # Ensure that this fails with "Access Denied"
     help
     # Ensure that md and rw are now prefixed with "-"

Signed-off-by: Craig Hesling <hesling@chromium.org>
Change-Id: I1202a118b836ff2a2acc986f6f3b0f6e07b36894