cr50: Add dummy U2F_ADPU vendor command.
This command is called by u2fd in M77 on startup,
and if a success response is not returned, u2fd
will not start up or send commands.

We need to support M77 in newer cr50 versions,
due to the enterprise rollback feature.

cr50 has been updated to not require these commands
to be sent, so the dummy command here does nothing
other than return success.

BUG=b:133275654
TEST=flash M77 and ToT cr50 with this change,
     check u2fd starts up and works
BRANCH=none

Signed-off-by: Louis Collard <louiscollard@chromium.org>
Change-Id: Ie1576f582df6c3a076a8b9a3cbc7101fb05204b3