fluffy: Fix voltage conversion accuracy
When converting the ADC reads into mV, I had used the wrong value which
was leading to some errors that would grow with the voltage.  This
commit simply uses the right values and attempts to gain more accuracy
by increasing the number of significant figures.

BUG=b:140393065
BRANCH=None
TEST=Flash fluffy; measure the VBUS voltage with a multimeter, verify
that the fluffy reported value is now closer to the multimeter.

Change-Id: I9dfd749eb533d96c8bc45454f968d7bcb095b9ef
Signed-off-by: Aseda Aboagye <aaboagye@google.com>