common: add __for_testing attribute
For functions which should only be used in test builds, this lets us
add __for_testing to the prototype, causing an error if used outside
of tests.

Example error output:

test.c: In function ‘main’:
test.c:17:5: error: call to ‘foo’ declared with attribute error:
             This function should only be used by tests
    foo();
    ^~~~~

This will not cause errors if the usage is guarded by a disabled
IS_ENABLED (or any other form of optimized-away branch).

BUG=none
BRANCH=none
TEST=see above

Change-Id: I64115fd9e7940de9b10063a46548e8d00033d1d3
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>