driver: lis2dw12/lis2dwl: fix wrong __fls usage
__fls(n) is defined as (31 - __builtin_clz(n)) in include/common.h.
Because of the definition, n can't be 0. When n is 0, __fls(0)
will be -1 and it is a wrong result.

Both LIS2DW12_ODR_TO_REG and LIS2DW12_ODR_TO_NORMALIZE can
receive a parameter which is less than LIS2DW12_ODR_MIN_VAL, so
__fls(0) will be used when runtime. That will give a wrong result.
This CL will fix it.

BUG=b:143242489
BRANCH=none
TEST=add debug output in driver, ensure LIS2DW12_ODR_TO_REG and
     LIS2DW12_ODR_TO_NORMALIZE give expected output.
     run 'suspend_stress_test --suspend_min=10 --wake_min=10 \
     --count=2500' on DUTs, test pass and no EC crash.

Change-Id: I46febb602b47624ba5d0106abaedd34a23ebe96f
Signed-off-by: Paul Ma <magf@bitland.corp-partner.google.com>