cr50: DO NOT SUBMIT - Allow U2F Sign without presence
Currently the signing function will sign any hash, with
no knowledge of the message. Given that this means it
could be signing a message that is a U2F authenticate
response that indicates presence was asserted, we must
always assert presence when signing.

This change illustrates one way that signing without
presence could be implemented, by verifying that the
message being signed is not a U2F message that indicates
presence was asserted.

BUG=b:140531796
TEST=build
BRANCH=none

Change-Id: I254fcaf54b0b8ebd1c6f908658fd0d14a4960c6c
Signed-off-by: Louis Collard <louiscollard@chromium.org>