Update patch set 1
Patch Set 1:

Really? A different SKU of the same board (not baseboard) would drop a USB port? That seems like a relatively big change. It seems that the semantics of the config option are changing and I don't think a mass find/replace is sufficient. You might have tasks trying to access objects that aren't quite applicable then.

Patch-set: 1
CC: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>