atlas: initialize max charge current
the isl923x is strapped to initialize the charge current to 3A.
however, its default max charge current limit is 3.072.  when the
charge current exceeds the current limit, the charger asserts PROCHOT
which means the AP gets throttled to 400MHz until the charge_state
machine updates the current limit.

on an unlocked system, we don't change the charge limit from its
default, so we never apply the 5% derating needed to avoid the isl923x
from over-currenting the charger.

the solution is to over-ride the 3A strapping of the isl923x by
appling a 5% derated current request early when we boot up.

BUG=b:141533503
BRANCH=none
TEST=atlas no longer boots into PROCHOT on 5v3a charger

Change-Id: Idba55edf7b1c0eec36b6583aa0b276c3cb1f0c89
Signed-off-by: Caveh Jalali <caveh@chromium.org>