Remove GPIOs for eSPI VW sleep signals
Remove non-interrupt GPIO configurations for SLP_S3_L and SLP_S4_L for
dragonegg and tglrvpu_ite. The GPIO names thus defined are only
referenced in the rest of the codebase if virtual wires are not enabled
for those signals, in which case the GPIOs will be configured as
interrupts.

BUG=b:143288478,b:139553375
TEST=Build dragonegg and tglrvpu_ite
BRANCH=none

Change-Id: I7b59bb315e333ab62257f206159d95a6d3c6a0ca
Signed-off-by: Abe Levkoy <alevkoy@chromium.org>