DO NOT SUBMIT: signal cr50 on fpsensor touched
This sample change uses Pin 0.3 as a new GPIO to connect to USER_PRES_L
on FPMCU, and toggle this line when fpsensor is touched in encoll or
match mode, in order to signal cr50 that the fingerprint sensor
is touched (in enroll or match mode). Also let cr50 read this line on
power button press.

TEST=observed reading toggles between 0 and 1 on cr50 console as I
	touch the fingerprint sensor in match mode and press power button.
	(Note: If the device is not in fingerprint enroll or match mode,
	this will not work.)

Change-Id: Ie616361e6787f4b090f8fe27d01be4d6da9fef9f
Signed-off-by: Yicheng Li <yichengli@chromium.org>