battery: keep battery awake
There is a known issue with TI battery fuel gauges that can cause the
battery to cutoff if the battery is sleeping. The TI fuel gauge will not
go to sleep if there has been i2c communication in the last 30 seconds,
so change our max polling frequency to 10 seconds.

BRANCH=octopus, and any other project with TI fuel gauge
BUG=b:133375756,chromium:1017366,chromium:1009611
TEST=keep battery awake greatly reduces sporadic battery cutoff in
manual tests.

Change-Id: Ib66a2ab03f9ed1363be5899d8bbfdc42cb041103
Signed-off-by: Jett Rink <jettrink@chromium.org>