fmap: Remove CONFIG_RO_STORAGE_OFF from EC_RO offset
The valud CONFIG_RO_STORAGE_OFF was being used to define the
area_offset for EC_RO. This means that when flashing EC_RO from the
DUT, the header region is not being written. This is causing issues
when doing RO updates (like in dogfood programs) as the full data
section is not always written.

This CL removes CONFIG_RO_STORAGE_OFF from the area_offset field of
EC_RO so that both the header and RO image are written.

BUG=b:142907781
BRANCH=None
TEST=Tested with a script that mimics auto update by writing an onlder
version of FW to RO, then uses the DUT to flash just the RO with a new
version and then compares the RO that was read to what was supposed to
be written. This script would always fail before, but with this CL now
passes multiple iterations.

Change-Id: I3d7af7de0dc1dfd24d6d304a8271290e6c6eb94e
Signed-off-by: Scott Collyer <scollyer@google.com>