Update patch set 1
Patch Set 1: Verified+1 Auto-Submit+1

this is the re-upload of

https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1670653

with a new change-id.  the original change-id started as a cherry-pick
from the atlas firmware branch, but this patch changed enough that it
should be submitted as a new change.  i'm going to revert the original
patch on the atlas firmware branch and replace it with this one.

Patch-set: 1
Reviewer: Gerrit User 1106949 <1106949@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1190764 <1190764@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Verified=+1
Label: Auto-Submit=+1
Work-in-progress: false