stm32f4: Add UNUSED declaration for gpio.inc
This new gpio.inc macros is intended to mark pins that are not connected
on a given board. This allows the chip implementation to configure them
for the lowest power configuration.

A reference implementation is provided for the STM32F4 chip family.
In this case, ST recommends that you configure the pin as analog-in to
disable the schmitt trigger logic.
This codified the work done in crrev.com/c/1883127 .

BRANCH=nocturne,hatch
BUG=b:130561737
TEST=make buildall -j
TEST=make BOARD=bloonchipper
     # Connect dragonclaw dev board over servo micro
     sudo servod --board=bloonchipper --config bloonchipper_rev0.1.xml &
     ./util/flash_ec --board=bloonchipper
     minicom -D $(dut-control raw_fpmcu_uart_pty | cut -d: -f2)
     > sysinfo
     # Check that we are in RW

Change-Id: I444b868b277b016a896e410dccae84429b68839e
Signed-off-by: Craig Hesling <hesling@chromium.org>