stm32f4: Add UNUSED declaration for gpio.inc
This new gpio.inc macros is intended to mark pins that are not connected
on a given board. This allows the chip implementation to configure them
for the lowest power configuration.

This CL brings no functional change.

A reference implementation is provided for the STM32F4 chip family.
In this case, ST recommends that you configure the pin as analog-in to
disable the schmitt trigger logic.

BRANCH=nocturne,hatch
BUG=b:130561737
TEST=make buildall -j

Change-Id: I0bc0a63401ae8f3bba4108b5b9f9ced26785f2bc
Signed-off-by: Craig Hesling <hesling@chromium.org>