stm32: Add UNUSED_IO declaration for gpio.inc
This new gpio.inc macros is intended to mark pins that are not connected
on a given board. This allows the chip implementation to configure them
for the lowest power configuration.

A reference implementation is provided for the STM32F4 chip family.
In this case, ST recommends you configure the pin as analog-in to
disable the schmitt trigger logic.
This codified the work done in crrev.com/c/1883127 .

BRANCH=nocturne,hatch
BUG=b:130561737
TEST=make buildall -j
TEST=make BOARD=hatch_fp
     # Flash bloonchipper to dragonclaw
     # Ensure it boots

Change-Id: I444b868b277b016a896e410dccae84429b68839e
Signed-off-by: Craig Hesling <hesling@chromium.org>