board_id: when checking headers ignore the RLZ if its erased
We will be able to set the board id flags without setting the RLZ. If
only flags are set, then check the flags. If the RLZ is set, also check
the RLZ.

BUG=b:143649068
BRANCH=cr50
TEST=set flags to 0x3f80. Try to update to a ZZAF:0:0:0 image. Make
sure it isn't rejected with RLZ mismatch. Try to update to a prepvt
image. Make sure it's rejected.

Change-Id: Ie0efdd7b1b6d76f385688f75c0765c08cab3755c
Signed-off-by: Mary Ruthven <mruthven@chromium.org>