board_id: clean up/enhancements
Add type_inv to the bid command output. In the bid output you can't tell
the difference between a board with RLZ 0xffffffff and a empty RLZ.
Change the command output to show type and type_inv, so we can tell the
difference.

Remove unused clear_flag parameter

BUG=b:143649068
BRANCH=cr50
TEST=run 'bid'
Change-Id: I13b6ba472010fdf85f94cb4015a9bbc48531973d
Signed-off-by: Mary Ruthven <mruthven@chromium.org>