Update patch set 1
Patch Set 1:

This is a bit of an experimental change for my first refactor of existing EC code.
The aim is simple, that the fans[] config was non-const, but only needed that for 2 boards.
I would like some guidance about this approach, and whether this is a reasonable thing to do.
Some more complicated options are possible such as using a board specific function to return the config for
those boards that require it. It's really just to start to get a handle on the way things are done in the EC.

Patch-set: 1
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1361259 <1361259@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>