stm32f4: Implement UNCONNECTED pins
This CL adds support for UNCONNECTED pins to the STM32F4 family
of chips. This places the unconnected/unused pins in the lowest
power state.

In this case, ST recommends that you configure the pin as analog-in to
disable the schmitt trigger logic.
This codified the work done in crrev.com/c/1883127 .

The CL should have no functional change for boards that have not
specified UNCONNECTED pins in their gpio.inc.

BRANCH=nocturne,hatch
BUG=b:130561737
TEST=make buildall -j

Change-Id: I444b868b277b016a896e410dccae84429b68839e
Signed-off-by: Craig Hesling <hesling@chromium.org>