stm32f4: Enable UNCONNECTED pins for hatch_fp
This CL adds support for UNCONNECTED pins to the STM32F4 family
of chips and enables it's use for hatch_fp. This places the unused
pins in the lowest power state for idle pins.

In this case, ST recommends that you configure the pin as analog-in to
disable the schmitt trigger logic.
This codified the work done in crrev.com/c/1883127 .

BRANCH=nocturne,hatch
BUG=b:130561737
TEST=make buildall -j
TEST=make BOARD=bloonchipper
     # Connect dragonclaw dev board over servo micro
     sudo servod --board=bloonchipper --config bloonchipper_rev0.1.xml &
     ./util/flash_ec --board=bloonchipper
     minicom -D $(dut-control raw_fpmcu_uart_pty | cut -d: -f2)
     > sysinfo
     # Check that we are in RW

Change-Id: I444b868b277b016a896e410dccae84429b68839e
Signed-off-by: Craig Hesling <hesling@chromium.org>