i2c: i2cxfer added wmod
We have a device that does not allow register offset into
the device.  When you write, you have to perform a block write
of the register values starting at 0.

    i2cxfer wmod 0 0x57 3 0xC2

This will read all 32 register values, modify the results of
the read at buff[3] to become 0xC2 and then will perform a
write of all 32 register values back out.

BUG=b:143775009
BRANCH=none
TEST=try to write a byte in the PI3DPX1207 and see it is written

Change-Id: I01e1d3c55c656ce5e18cae60cd9cc74c8bc6a3d3
Signed-off-by: Denis Brockus <dbrockus@chromium.org>