i2c [DEBUG ONLY]: i2cxfer added rb/wb [DO NOT MERGE]
We have a device that does not allow register offset into
the device.  When you write, you have to perform a block write
of the register values starting at 0.  An example of reading
and writing

    i2cxfer rb 0 0x57 3
    i2cxfer wb 0 0x57 3 0xC2

It looks like the normal r/w modes of this command but
internally it reads the registers starting at 0 to
mimic this functionality

BUG=b:143775009
BRANCH=none
TEST=try to write a byte in the PI3DPX1207 and see it is written

Change-Id: I01e1d3c55c656ce5e18cae60cd9cc74c8bc6a3d3
Signed-off-by: Denis Brockus <dbrockus@chromium.org>