baseboard/kukui: dynamic adjuct input current by thermal budget
Limit the charger input current if we are overheat,
and loose the limit if the thermal in control.

kukui/krane/kodama are suffering from thermal issue (>100 celsius
degree in charger junction) and the solution was limit the system
can only sink 5V/2A in S0.  However, this cause slowing charging
as well.  We'd like to loose the 5V/2A charging limit and controls
the thermal by restricting the charger input current.

This CL only applies the limitation when PD voltage > 5V, because
we don't see the overheat problem when krane charged with 5V.

TEST=on krane, ensure the temperature of junction is always under
     84 celsius degree; ensure the curve of battery current at
     charging is a smooth curve when PD voltage switching from high
     to low.
BUG=b:141903096 b:144073892
BRANCH=kukui

Change-Id: I9657f5a04897010d7018f910cb77b210147d8356
Signed-off-by: Yilun Lin <yllin@chromium.org>
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>