TCPMv2: Enable TCPC low power when port is looking for a connection
When the port is in a state where it is looking for a connection,
to save power, we should put the TCPC in its low power mode and
enable auto toggling. Low power mode can happen when DRP auto
toggling, acting as a SNK only, or acting as a SRC only.

BUG=chromium:1022217
BRANCH=none
TEST=make -j buildall
	manual tests:
	1: (S0) Nothing plugged in, port is drp and low power mode
	2: (S5/S3/S0ix) Port is SNK only, and low power with nothing plugged in
	3: (S3/S0ix) If TypeC sink was previously plugged in, port remains powered
	4: (S5/S3/S0ix) TypeC source is recognized
	5: (S3->S0) TypeC sink plugged in, port is powered when S0 is reached

Change-Id: I7e853d05e0ece1f6b3031f17a18fcbf0d9a15a51
Signed-off-by: Sam Hurst <shurst@google.com>