helios: Fix battery discharge FET access config setting
The batteries for helios specified register 0x44 as the register address
that contains the status of the discahrge FET. That is correct for the
TI fuel gauge. However, address 0x44 is in the manufacturer access
region and therfore .mfgacc_support field needs to be set. However,
register 0x0 mirrors the lower 16 bits of mfgacc reg 0x44, so an
alternate solution is to change the reg_addr field to 0x0 and keep the
regular smart battery register access method.

In addition, this CL also removes dependency on the charge FET for
battery disconnect state as that really only depends on the discharge
FET being active so the battery can supply power to the system.

BUG=b:136970148
BRANCH=None
TEST=Put helios in ship mode via 'cutoff' on EC console. Then after
waiting 30 seconds apply external AC power. Verfied that the discharge
FET is initially disabled, then gets enabled as the battery wakes up.

2019-11-08 14:48:54 [3.428296 Batt disconnected: reg 0x6301
2019-11-08 14:48:54 [3.428843 found battery in disconnect state]
2019-11-08 14:48:54 [3.678447 Batt disconnected: reg 0x6301
2019-11-08 14:48:54 [3.678995 found battery in disconnect state]
2019-11-08 14:48:55 [4.427645 charge_manager_leave_safe_mode()]
2019-11-08 14:48:55 [4.434650 Battery 99% (Display 100.0 %) / ??h:??
to full]

Prior to this fix this register was reading back 0x16 always.

Change-Id: I8fbcdeb55f692796ced034c3c95db94fafdc908e
Signed-off-by: Scott Collyer <scollyer@google.com>